Fix some issues on Windows with the latest changes in trunk.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@761421 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/core/ActiveMQConsumer.h
@@ -111,7 +111,7 @@ namespace core{
         /**
          * Time to wait before restarting delivery of rollback messages.
          */
-        long redeliveryDelay;
+        long long redeliveryDelay;
 
         /**
          * Has the Synchronization been added for this transaction
